% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes_1987.R
\docType{data}
\name{aes_1987}
\alias{aes_1987}
\title{Australian Election Study 1987}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1830 rows and 195 columns.}
\source{
\url{http://www.australianelectionstudy.org/voter_studies.html}
}
\usage{
aes_1987
}
\description{
\itemize{
  \item Total Sample: 3,061
  \item Valid responses: 1,825
  \item Effective response: 62.8\%
}
}
\details{
The study has two goals. The first is to continue the broad line of enquiry established by the
1967 and 1979 Australian National Political Attitudes surveys so that patterns of stability and
change in the political attitudes and behaviour of the Australian electorate can be traced over
two decades. The second is to assess the electoral impact of forces specific to this election in
order better to understand its outcome. Respondents were asked which party they voted for in the
1987 and 1984 Federal elections; when they had decided how to vote and whether they might have
changed their vote; party identification; interest in politics; the importance of a number of
issues and the media in deciding their vote; left-right positions for their own views and those
of the major parties in the 1987 and 1984 elections; feelings about the party leaders indicated
by ratings from 0 to 10, their qualities and effectiveness as prime minister.

A section on election issues covered perceptions of Australia's current economic situation and
the extent of the effect of government policies on the economy; government spending;
privatisation; the role of trade unions and big business; uranium mining; law and order; Asian
migration; marijuana; abortion; aborigines; testing for AIDS; censorship; homosexuality;
opportunities for women. Other questions examined trust in government; political goals; and forms
of political action.

Background information included level of schooling; highest level of qualification; current
activity; occupation; employment sector; supervision; union membership; self assessed social
class; sex; age; length of current residence; country of birth of self and parents; year arrived
in Australia; father's occupation; parents' schooling and political preference; marital status;
number of children; partner's schooling, occupation, employment sector, supervision, union
membership and political preference; religion and attendance at religious services.

\strong{Useful links}

\itemize{
  \item \href{http://ada.edu.au/ADAData/codebooks/ADA.CODEBOOK.00445.pdf.zip}{Original codebook}
  \item \href{http://ada.edu.au/ADAData/questionnaires/ADA.QUESTIONNAIRE.00445.pdf.zip}{Original questionanaire}
}
}
\section{Variables}{


\describe{
	 \item{ID}{IDENTIFICATION NUMBER}
	 \item{CARD1}{CARD NUMBER}
}

\subsection{SECTION A: THE FEDERAL ELECTION}{
\describe{
  \item{Q1}{CARE WHICH PARTY WINS ELECTION}
	 \item{Q2TV}{FOLLOWED ELECTION NEWS ON TV}
	 \item{Q2PAPERS}{FOLLOWED ELECTION NEWS IN NEWSPAPERS}
	 \item{Q2RADIO}{FOLLOWED ELECTION NEWS ON RADIO}
	 \item{Q3TV}{TV IMPORTANT IN VOTE DECISION}
	 \item{Q3PAPERS}{PAPERS IMPORTANT IN VOTE DECISION}
	 \item{Q3RADIO}{RADIO IMPORTANT IN VOTE DECISION}
	 \item{Q4WHEN}{DECIDED HOW TO VOTE}
	 \item{Q5REPS}{VOTE IN HOUSE OF REPRESENTATIVES}
	 \item{Q5SENATE}{VOTE IN SENATE}
	 \item{Q6}{REASON CHOSE PARTY FOR HOUSE OF REPS}
	 \item{Q7}{OTHER CHOICE FOR HOUSE OF REPS}
	 \item{Q8REPS}{1984 VOTE IN HOUSE OF REPRESENTATIVES}
	 \item{Q8SENATE}{1984 VOTE IN SENATE}
	 \item{Q9}{VOTING STABILITY}
	 \item{Q1OUNION}{IMPORTANT ISSUES - TRADE iJNTON}
	 \item{Q1OUNEMP}{IMPORTANT ISSUES}
	 \item{Q1OINFLT}{IMPORTANT ISSUES - INFLATION}
	 \item{Q1OINTR}{IMPORTANT ISSUES - INTEREST RATES}
	 \item{Q1OHEALT}{IMPORTANT ISSUES - HEALTH CARE}
	 \item{Q1OLIVST}{IMPORTANT ISSUES - LIVING STANDARDS}
	 \item{Q1OGOVSP}{IMPORTANT ISSUES - GOVERNMENT SPENDING }
	 \item{Q1OTAXES}{IMPORTANT ISSUES}
	 \item{Q1ORURAL}{IMPORTANT ISSUES - THE RURAL CRISIS}
	 \item{Q1OENVIR}{IMPORTANT ISSUES - THE ENVIRONMENT}
	 \item{Q11}{COMPULSORY VOTING}
	 \item{Q12}{PARTY IDENTIFICATION}
	 \item{Q13}{STRENGTH OF PARTY ID}
	 \item{Q14}{INTEREST IN POLITICS}
	 \item{Q15}{RESPONDENT’S LEFT-RIGHT POSITION}
	 \item{Q16LIBS}{LIBERAL’S LEFT-RIGHT POSITION}
	 \item{Q16LABOR}{LABOR’S LEFT-RIGHT POSITION}
	 \item{Q16NATS}{NATIONAL’S LEFT-RIGHT POSITION}
	 \item{Q17}{RESPONDENT’S 1984 LEFT-RIGHT POSITION}
	 \item{Q18LIBS}{LIBERAL’S 1984 LEFT-RIGHT POSITION}
	 \item{Q18LABOR}{LABOR’S 1984 LEFT-RIGHT POSITION}
	 \item{Q18NATS}{NATIONAL’S 1984 LEFT-RIGHT POSITION}
}
}

\subsection{SECTION B POLITICAL LEADERS}{
\describe{
  \item{Q19BOBH}{FEELINGS ABOUT BOB HAWKE}
	 \item{Q19JOHNH}{FEELINGS ABOUT JOHN HOWARD}
	 \item{Q19IANS}{FEELINGS ABOUT IAN SINCLAIR}
	 \item{Q19JANH}{FEELINGS ABOUT JANINE HAINES}
	 \item{Q19ANDYP}{FEELINGS ABOUT ANDREW PEACOCK}
	 \item{Q19PAULK}{FEELINGS ABOUT PAUL KEATING}
	 \item{Q19JOHBP}{FEELINGS ABOUT JOH BJELKE PETERSEN}
	 \item{Q2OBOBH}{HAWKE EFFECTIVE AS P.M.}
	 \item{Q2OJOHNH}{HOWARD EFFECTIVE AS P.M.}
	 \item{Q2OIANS}{SINCLAIR EFFECTIVE AS P.M}
	 \item{Q21BH1}{HAWKE CARING}
	 \item{Q21BH2}{HAWKE DETERMINED}
	 \item{Q21BH3}{HAWKE SHREWD}
	 \item{Q21BH4}{HAWKE LIKEABLE AS A PERSON}
	 \item{Q21BH5}{HAWKE TOUGH}
	 \item{Q21BH6}{HAWKE LISTENS TO REASON}
	 \item{Q21BH7}{HAWKE DECISIVE}
	 \item{Q21BH8}{HAWKE STICKS TO PRINCIPLES }
	 \item{Q21BH9}{HAWKE SINCERE}
	 \item{CARD2}{CARD NUMBER 2}
	 \item{Q21JH1}{HOWARD CARING}
	 \item{Q21JH2}{HOWARD DETERMINED}
	 \item{Q21JH3}{HOWARD SHREWD}
	 \item{Q21JH4}{HOWARD LIKEABLE AS A PERSON}
	 \item{Q21JH5}{HOWARD TOUGH}
	 \item{Q21JH6}{HOWARD LISTENS TO REASON}
	 \item{Q213H7}{HOWARD DECISIVE}
	 \item{Q21JH8}{HOWARD STICKS TO PRINCIPLES}
	 \item{Q21JH9}{HOWARD SINCERE}
	 \item{Q21IS1}{SINCLAIR CARING}
	 \item{Q211S2}{SINCLAIR DETERMINED}
	 \item{Q21IS3}{SINCLAIR SHREWD}
	 \item{Q21IS4}{SINCLAIR LIKEABLE AS A PERSON}
	 \item{Q21ISS}{SINCLAIR TOUGH}
	 \item{Q21IS6}{SINCLAIR LISTENS TO REASON}
	 \item{Q211S7}{SINCLAIR DECISIVE}
	 \item{Q21IS8}{SINCLAIR STICKS TO PRINCIPLES}
	 \item{Q21IS9}{SINCLAIR SINCERE}
	 \item{Q22UNITE}{LEADER ABLE TO UNITE THE NATION}
	 \item{Q22DQER}{LEADER LIKELY TO GET THINGS DONE}
	 \item{Q22OSEAS}{LEADER LIKELY TO IMPROVE STANDING ABROAD}
	 \item{Q22CNCRN}{LEADER MOST CONCERNED FOR ALL}
	 \item{Q22TEAM}{LEADER GET MOST OUT OF A TEAM}
	 \item{Q22TOUCH}{LEADER IN TOUCH WITH PEOPLE’S PROBLEMS}
	 \item{Q22ECON}{LEADER LIKELY TO SOLVE ECONOMIC PROBLEMS}
	 \item{Q23LABOR}{LABOR PARTY EXTREME - MODERATE}
	 \item{Q23LIB}{LIBERAL PARTY EXTREME - MODERATE}
	 \item{Q23NAT}{NATIONAL PARTY EXTREME - MODERATE}
	 \item{Q24LABOR}{LABOR PARTY GOOD FOR}
	 \item{Q24LIB}{LIBERAL PARTY GOOD FOR}
	 \item{Q24NAT}{NATIONAL PARTY GOOD FOR}
	 \item{Q25LABOR}{LABOR PARTY UNITED - DIVIDED}
	 \item{Q25LIB}{LIBERAL PARTY UNITED - DIVIDED}
	 \item{Q25NAT}{NATIONAL PARTY UNITED - DIVIDED}
}
}

\subsection{SECTION C: ELECTION ISSUES}{
\strong{THE ECONOMY}
  \describe{
  \item{Q26}{HOUSEHOLD ECONOMY NOW CF 1 YEAR AGO}
	 \item{Q27}{COUNTRY ECONOMY CF 1 YEAR AGO}
	 \item{Q28HHOLD}{EFFECT OF GOVT POLICY ON H’HOLD ECONOMY}
	 \item{Q28CNTRY}{EFFECT OF GOVT POLICY ON COUNTRY ECONOMY}
	 \item{Q29}{ANGER OVER GOVT HANDLING OF ECONOMY}
	 \item{Q30}{GOVT POLICY IMPROVE OR WORSEN ECONOMY}
	 \item{Q31}{CHOICE BETWEEN TAXES AND SOCIAL SERVICES}
	 \item{Q32UNION}{POWER OF TRADE UNIONS}
	 \item{Q32BUSNS}{POWER OF BIG BUSINESS}
	 \item{Q33UNION}{TRADE UNION LEADERS USE OF POWER}
	 \item{Q33BUSNS}{BIG BUSINESS LEADERS USE OF POWER}
	 \item{Q34TAX}{HIGH TAX DISINCENTIVE}
	 \item{Q34EQUAL}{INCOME AND WEALTH REDISTRIBUTION}
	 \item{Q34WLFAR}{RELIANCE ON GOVERNMENT HANDOUTS}
	 \item{Q34PVRTY}{SPEND TO REDUCE POVERTY}
	 \item{Q3SREGTU}{LAWS TO REGULATE TRADE UNIONS}
	 \item{Q3SWKRUN}{GIVE WORKERS MORE SAY AT WORK}
	 \item{Q35PRFIT}{PRIVATE INDUSTRY KEEP PROFITS}
	 \item{Q3SPMED}{ENCOURAGE PRIVATE MEDICINE}
  }
\strong{NON-ECONOMIC ISSUES}
\describe{
 \item{Q37}{ATTITUDE TO URANIUM MINING}
	 \item{Q38}{LEGALISE MARIJUANA}
	 \item{Q39}{ABORTION}
	 \item{Q40}{ROYALTY}
	 \item{Q41}{AUSTRALIA’S LINKS WITH THE U.S.}
	 \item{Q42}{CHANGE AUSTRALIA’S FLAG}
	 \item{Q43}{CRIME CRIME AND UNEMPLOYMENT}
	 \item{Q43POLIC}{MORE POWER TO POLICE}
	 \item{Q43DPEN}{BRING BACK THE DEATH PENALTY}
	 \item{Q43SNTC}{STIFFER SENTENCES FOR BREAKING THE LAW 46C}
	 \item{Q44}{ASIAN IMMIGRATION POLICY}
	 \item{Q45}{BENEFITS FOR ABORIGINES}
	 \item{Q46IMGTS}{AIDS TESTING - IMMIGRANTS}
	 \item{Q46HOMO}{AIDS TEST - HOMOSEXUAL}
	 \item{Q46FORCS}{AIDS TEST - THE DEFENCE FORCES}
	 \item{Q46JAILS}{AIDS TEST - PRISONERS IN JAIL}
	 \item{Q46TOURS}{AIDS TEST - TOURISTS ENTERING AUSTRALIA}
	 \item{Q46MUMS}{AIDS TEST - PREGNANT WOMEN}
	 \item{Q47}{CENSORSHIP OF BOOKS AND FILMS}
	 \item{Q48}{SYMPATHY FOR OR AGAINST STRIKERS}
	 \item{Q49}{HOMOSEXUAL ACTS BETWEEN CONSENTING ADULT}
	 \item{CARD3}{CARD NUMBER 3}
	 \item{Q50}{JOB OPPORTUNITIES FOR WOMEN}
	 \item{Q51EQOP}{EQUAL OPPORTUNITY FOR WOMEN}
	 \item{QS1PORN}{THE RIGHT TO SHOW NUDITY AND SEX}
	 \item{QS1ENVRN}{PROTECTION OF THE ENVIRONMENT}
	 \item{Q51ETHNC}{IMMIGRANTS KEEPING THEIR ETHNICITY}
	 \item{QS1ABOLR}{TRANSFERRING LAND RIGHTS TO ABORIGINES}
	 }
}

\subsection{SECTION D: SOCIAL AND POLITICAL GOALS}{
\describe{
  \item{Q52}{MOST IMPORTANT GOAL FOR AUSTRALIA}
	 \item{Q53}{TRUST IN FEDERAL GOVERNMENT}
	 \item{Q54}{FEDERAL GOVERNMENT RUN FOR BENEFIT OF}
	 \item{Q55}{FEDERAL GOVERNMENT WASTE MONEY}
	 \item{Q56}{FEDERAL GOVERNMENT DISHONEST}
	 \item{Q57}{FEELINGS ABOUT VOTING}
	 \item{Q58}{NEED ELECTION CAMPAIGN OR NOT}
	 \item{QS9ACT1}{SIGNING A PETITION}
	 \item{Q59ACT2}{JOINING IN BOYCOTTS}
	 \item{Q59ACT3}{ATTENDING LAWFUL DEMONSTRATIONS}
	 \item{Q59ACT4}{JOINING UNOFFICIAL STRIKES}
	 \item{Q59ACT5}{OCCUPYING BUILDINGS AND FACTORIES}
	 \item{Q59ACT6}{DAMAGING THINGS}
	 \item{Q59ACT7}{USING PERSONAL VIOLENCE}
}
}

\subsection{SECTION E: EDUCATION AND WORK}{
\strong{EDUCATION}
\describe{
	 \item{Q60}{HIGHEST GRADE OF SCHOOL COMPLETED}
	 \item{Q61}{AGE LEFT SCHOOL}
	 \item{Q62}{TYPE OF SECONDARY SCHOOL}
	 \item{Q63}{YEARS OF TERTIARY EDUCATION}
	 \item{Q64}{LEVEL OF TERTIARY QUALIFICATION OBTAINED}
}
\strong{WORK}
\describe{
  \item{Q65}{ACTIVITY LAST WEEK}
	 \item{Q66}{OCCUPATION}
	 \item{Q66}{D EMPLO~R}
	 \item{Q67}{DOE~R15i~’uNDENT SUPERVISE}
	 \item{Q68}{TRADE UNION MEMBERSHIP}
	 \item{Q69}{YEARS BELONGED TO TRADE UNION}
	 \item{Q70}{SOCIAL CLASS - SELF PERCEPTION}
}
}

\subsection{SECTION F: PERSONAL BACKGROUND}{
\describe{
	 \item{Q71}{SEX}
	 \item{Q72}{AREA OF RESIDENCE}
	 \item{Q73}{YEARS LIVED IN THIS AREA}
	 \item{Q74}{YEAR OF BIRTH}
	 \item{Q75RCOB}{RESPONDENT’S COUNTRY OF BIRTH}
	 \item{Q7SMCOB}{MOTHER’S COUNTRY OF BIRTH}
	 \item{Q75FCOB}{FATHER’S COUNTRY OF BIRTH}
	 \item{Q76}{YEAR CAME TO AUSTRALIA}
	 \item{Q77}{FATHER’S OCCUPATION WHEN R WAS 14 YEARS}
	 \item{Q7SF}{AGE FATHER LEFT SCHOOL}
	 \item{Q78M}{AGE MOTHER LEFT SCHOOL}
	 \item{Q79F}{FATHER’S POLITICAL PREFERENCE}
	 \item{Q79M}{MOTHER’S POLITICAL PREFERENCE}
	 \item{Q80}{MARITAL STATUS}
	 \item{Q81}{NUMBER OF CHILDREN}
	 \item{Q82}{AGE PARTNER LEFT SCHOOL}
	 \item{Q83}{IS PARTNER WORKING NOW}
	 \item{Q84}{PARTNER’S OCCUPATION}
	 \item{CARD4}{CARD NUMBER 4}
	 \item{Q85}{PARTNER’S EMPLOYER}
	 \item{Q86}{PARTNER A SUPERVISOR}
	 \item{Q87}{PARTNER MEMBER OF TRADE UNION}
	 \item{Q88}{PARTNER’S POLITICAL PREFERENCE}
	 \item{Q89}{RELIGIOUS DENOMINATION}
	 \item{Q90}{CHURCH ATTENDANCE}
	 \item{STATE}{STATE OF RESIDENCE}
	 \item{ELECDIV}{ELECTORAL DIVISION}
	 \item{POSTCODE}{POSTCODE OF RESIDENCE}
}
}
}

\keyword{datasets}
