% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes_1990.R
\docType{data}
\name{aes_1990}
\alias{aes_1990}
\title{Australian Election Study 1990}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2037 rows and 236 columns.}
\source{
\url{http://www.australianelectionstudy.org/voter_studies.html}
}
\usage{
aes_1990
}
\description{
\itemize{
  \item Total sample: 3,606
  \item Valid responses: 2,020
  \item Effective response rate: 58.0
}
}
\details{
The Australian Election Study is the second in the series initiated in 1987, and builds on the
data collected earlier in the 1967 and 1979 Australian Political Attitudes Surveys. The series
provides a long-term perspective on stability and change in the political attitudes and behaviour
of the Australian electorate, and investigates the changing social bases of Australian politics
as the economy and society modernise and change character. In addition to these long-term
perspectives it examines the political issues prevalent in the election and assesses their
importance for the election result.

Two major themes are covered in the survey, the economy and the environment. Questions on the
economy include retrospective and prospective evaluations of the country's economy and the
individual's own personal household situation; wages and price controls; attitudes to particular
economic issues; and the position of the major parties and party leaders on the main economic
issues. Questions on the environment include attitudes to the protection of the environment and
to environmental issues in general; the role of industry and the government in ensuring
environmental protection; rankings of the most urgent environmental issues and the respondent's
assessment of their likely future importance; and attitudes toward environmental pressure and
interest groups, including current or likely membership.

As in the 1987 AES, a range of questions are included in the survey which cover election issues
and the respondent's assessment of the most effective party to handle them, as well as questions
on specific political issues such as government spending and the pilots' dispute. A section on
social policy includes questions on around 30 contemporary social and moral issues, including
abortion, Aboriginal rights, pornography, health care and many others. A section on industrial
policy focuses on attitudes towards the control of the economy, attitudes to trade unions and big
business, and to industrial conflict in general. In addition to the standard questions on vote in
the current election, recalled vote in the 1987 federal election, and party identification, the
survey contains an extensive section dealing with attitudes to the major parties and to the party
leaders. Finally, a range of demographic data are collected for each respondent and their spouse
or partner, if applicable.

\strong{Useful links}

\itemize{
   \item \href{http://ada.edu.au/ADAData/codebooks/ADA.CODEBOOK.00445.pdf.zip}{Original codebook}
   \item \href{http://ada.edu.au/ADAData/questionnaires/ADA.QUESTIONNAIRE.00445.pdf.zip}{Original questionnaire}
}
}
\section{Variables}{


Note: this documentation has been autogenerated from the dataset labels not the original codebook.

\describe{
 \item{id}{Respondent Id Number}
 \item{a1tv}{Followed election news on TV}
 \item{a1papers}{Followed election news in the newspapers}
 \item{a1radio}{Followed election news on the radio}
 \item{a2tv}{Watched campaign ads on TV}
 \item{a2papers}{Read campaign ads in the newspapers}
 \item{a2radio}{Listened campaign ads on the radio}
 \item{a3}{Notice taken of polls}
 \item{a4}{Interest in politics}
 \item{a5}{Watched televised debate}
 \item{a6}{Who performed better in televised debate}
 \item{a7}{Watched formal election policy launches}
 \item{a8}{Most impressive party launch}
 \item{a9ad}{Awareness of Aust Democrats campaign}
 \item{a9greens}{Awareness of Greens campaign}
 \item{b1}{Party identification}
 \item{b2}{Strength of party identification}
 \item{b3}{Left-right position}
 \item{b4}{When decided which way to vote}
 \item{b5}{Which party would win election}
 \item{b6reps}{Vote in House of Representatives}
 \item{b6senate}{Vote in Senate}
 \item{b7}{Thought of changing vote during campaign}
 \item{b8reps}{Vote in House of Representatives in 1987}
 \item{b8senate}{Vote in Senate in 1987}
 \item{b9}{Previous vote}
 \item{c1bh}{Feelings about Bob Hawke}
 \item{c1ap}{Feelings about Andrew Peacock}
 \item{c1cb}{Feelings about Charles Blunt}
 \item{c1janh}{Feelings about Janine Haines}
 \item{c1pk}{Feelings about Paul Keating}
 \item{c1johnh}{Feelings about John Howard}
 \item{c2}{Like about Bob Hawke}
 \item{c3}{Dont like about Bob Hawke}
 \item{c4}{Like about Andrew Peacock}
 \item{c5}{Dont like about Andrew Peacock}
 \item{c6alp}{ALP extreme or moderate}
 \item{c6lib}{Coalition extreme or moderate}
 \item{c7alp}{ALP capable of strong government or not}
 \item{c7lib}{Coalition capable of strong government o}
 \item{c8alp}{ALP united or divided}
 \item{c8lib}{Coalition united or divided}
 \item{c9alp}{ALP has clear or vague policies}
 \item{c9lib}{Coalition has clear or vague policies}
 \item{d1unem}{Party closest to own view unemployment}
 \item{d1inflat}{Party closest to own view on inflation}
 \item{d1health}{Party closest to own view on health}
 \item{d1wages}{Party closest to own view on wages}
 \item{d1intr}{Party closest to own view interest rates}
 \item{d1educ}{Party closest to own view on education}
 \item{d1envir}{Party closest to own view on environment}
 \item{d1tax}{Party closest to own view on taxation}
 \item{d1gspend}{Party closest to own view on government}
 \item{d2unem}{Importance of unemployment issue}
 \item{d2inflat}{Importance of inflation issue}
 \item{d2health}{Importance of health issue}
 \item{d2wages}{Importance of wages issue}
 \item{d2intr}{Importance of interest rates issue}
 \item{d2educ}{Importance of education issue}
 \item{d2envir}{Importance of environment issue}
 \item{d2tax}{Importance of taxation issue}
 \item{d2gspend}{Importance of government spending issue}
 \item{d2i1}{Issue of most concern}
 \item{d2i2}{Second issue of concern}
 \item{d3less1}{Top priority for less govt spending}
 \item{d3less2}{Second priority for less govt spending}
 \item{d3more1}{Top priority for more govt spending}
 \item{d3more2}{Second priority for more govt spending}
 \item{e1hhold}{Financial situation of hhold cf 1 year a}
 \item{e1cntry}{Economic situation of country cf 1 year}
 \item{e2hhold}{Effect of govt policies on hhold situati}
 \item{e2cntry}{Effect of govt policies on country situa}
 \item{e3hhold}{Financial situation of hhold cf 1 yr ahe}
 \item{e3cntry}{Economic situation of country cf 1 yr ah}
 \item{e4hhold}{Effect of govt policies on hhold situati}
 \item{e4cntry}{Effect of govt policies on country situa}
 \item{e5wages}{Control of wages by legislation}
 \item{e5prices}{Control of prices by legislation}
 \item{e5hlthed}{Reduce expenditure on health & education}
 \item{e5import}{Increasing controls on imports}
 \item{e5subpi}{Increase Govt subsidies for private indu}
 \item{e5defnce}{Reducing expenditure on defence}
 \item{e6tax}{High tax makes people less willing to wo}
 \item{e6equal}{Income & wealth should be redistributed}
 \item{e6pind}{Private industry keep more of its profit}
 \item{e6profit}{Modern society without profit motive}
 \item{e7resp}{Distribution of wealth - Rs position}
 \item{e7alp}{Distribution of wealth - ALP position}
 \item{e7lib}{Distribution of wealth - Coalition posit}
 \item{e7bh}{Distribution of wealth - Hawkes position}
 \item{e7ap}{Distribution of wealth - Peacocks positi}
 \item{f1resp}{Taxes v. social services - Rs position}
 \item{f1alp}{Taxes v. social services - ALP position}
 \item{f1lib}{Taxes v. social services - Coalition pos}
 \item{f1bh}{Taxes v. social services - Hawkes positi}
 \item{f1ap}{Taxes v. social services - Peacocks posi}
 \item{f2}{Private medical treatment}
 \item{f3}{Abortion}
 \item{f4}{Government help for Aborigines}
 \item{f5equop}{Equal opportunities for women}
 \item{f5nudsex}{Nudity and sex in films and magazines}
 \item{f5ethnid}{Migrants keeping their ethnic identity}
 \item{f5aborlr}{Transfer of land rights to Aborigines}
 \item{f5migeo}{Equal opportunities for migrants}
 \item{f5protes}{The right to protest and demonstrate}
 \item{f5immig}{No. of migrants allowed into Australia}
 \item{f6deathp}{Death penalty reintroduced for murder}
 \item{f6homlaw}{Stricter laws against homosexuality}
 \item{f6marij}{Decriminalise smoking of marijuana}
 \item{f6polpow}{The police should be given more power}
 \item{f6lawbrk}{Stiffer sentences if break the law}
 \item{f6hlthtx}{Increase taxes and spend more on health}
 \item{f6govhnd}{Too many rely on government handouts}
 \item{f6redpov}{More money spent on reducing poverty}
 \item{f6succes}{In Australian society anyone can succeed}
 \item{f6wifhom}{Husband bread-winner and wife home}
 \item{f6eductx}{Increase taxes & spend more on education}
 \item{f6unemp}{Most of the unemployed dont want to work}
 \item{f6govser}{Govt serves the interests of Australians}
 \item{f7smkgov}{Ban smoking: in government departments}
 \item{f7smkwrk}{Ban smoking: The workplace generally}
 \item{f7smkeat}{Ban smoking: Public eating places}
 \item{f7smkshp}{Ban smoking: Shops and supermarkets}
 \item{f7smkpub}{Ban smoking: All public places}
 \item{f8}{Whether currently a smoker}
 \item{f9aims1}{First choice Australian aims}
 \item{f9aims2}{Second choice Australian aims}
 \item{f10aim1}{First choice national aims}
 \item{f10aim2}{Second choice national aims}
 \item{g1resp}{Logging in N. Est. Forests - Rs position}
 \item{g1alp}{Logging in N. Est. Forests - ALP positio}
 \item{g1lib}{Logging in N. Est. Forests - Coalition p}
 \item{g1bh}{Logging in N. Est. Forests - Hawkes posi}
 \item{g1ap}{Logging in N. Est. Forests - Peacocks po}
 \item{g2}{Views on prices and the environment}
 \item{g3}{Views on taxes and the environment}
 \item{g4}{Nuclear power - danger or resource}
 \item{g5pollut}{Pollution urgent envir. concern}
 \item{g5pop}{Overpopulation urgent envir. concern}
 \item{g5waste}{Waste disposal urgent envir. concern}
 \item{g5uran}{Uranium mining urgent envir. concern}
 \item{g5log}{Logging urgent envir. concern}
 \item{g5wild}{Wildlife urgent envir. concern}
 \item{g5soil}{Soil urgent envir. concern}
 \item{g5ghouse}{Greenhouse urgent envir. concern}
 \item{g5urg1}{Most urgent envir. concern}
 \item{g5urg2}{Second most urgent envir. concern}
 \item{g6pollut}{Pollution worse in 10 years}
 \item{g6pop}{Overpopulation worse in 10 years}
 \item{g6waste}{Waste disposal worse in 10 years}
 \item{g6uran}{Uranium mining worse in 10 years}
 \item{g6log}{Logging worse in 10 years}
 \item{g6wild}{Wildlife worse in 10 years}
 \item{g6soil}{Soil degredation worse in 10 years}
 \item{g6ghouse}{Greenhouse worse in 10 years}
 \item{g7tu}{Trust Trade Unions}
 \item{g7envir}{Trust Environmental groups}
 \item{g7comgov}{Trust Commonwealth Government}
 \item{g7stgov}{Trust State Governments}
 \item{g7polpar}{Trust Political parties}
 \item{g7ind}{Trust Industry}
 \item{g7scitec}{Trust Scientists and technologists}
 \item{g7public}{Trust general public}
 \item{g8envir}{Approve environmental groups}
 \item{g8uran}{Approve anti-uranium mining groups}
 \item{g8war}{Approve anti-war and anti-nuclear groups}
 \item{g9envir}{Member of environment group}
 \item{g9uran}{Member of anti-uranium mining group}
 \item{g9war}{Member of anti-war and anti-nuclear grou}
 \item{g10nship}{Ban nuclear powered warships}
 \item{g10nengy}{Nuclear energy a necessity}
 \item{g10techg}{Good effects of technology outweigh bad}
 \item{g10scit}{Science & technology best hope for futur}
 \item{g10uran}{Australia should mine its uranium}
 \item{g10envir}{Our society is abusing the environment}
 \item{g10pop}{Approaching world overpopulation}
 \item{g10natok}{Other things more important than Nature}
 \item{g10presn}{Nature most precious in life}
 \item{g10prote}{Could do more to protect the environment}
 \item{g10govte}{Govt could do more to protect environmen}
 \item{h1}{Employee control of industry}
 \item{h2}{Sympathy for or against strikers}
 \item{h3}{Government handling of pilot dispute}
 \item{h4}{Coalition handling of pilot dispute}
 \item{h5jobs}{Govt ensure jobs for all who want work}
 \item{h5regtu}{Stricter regulation of trade unions}
 \item{h5strke}{Workers strike without good reason}
 \item{h5tupowr}{Trade union power}
 \item{h5buspwr}{Big business power}
 \item{h5tuprot}{Trade unions protect workers}
 \item{h5busben}{Big business exploits workers}
 \item{h5pvtis}{Privatisation of industry}
 \item{h5wrksay}{Workers say in workplace}
 \item{h5scabs}{Scab labour prohibited by law}
 \item{h5class}{No class differences in Australia}
 \item{i1}{Age left School}
 \item{i2}{Private or Non-private school education}
 \item{i3}{No. years of tertiary education}
 \item{i4}{Highest qual obtained since leaving scho}
 \item{i5}{Occupation in past week}
 \item{i6}{Occupation}
 \item{i6d}{Employer}
 \item{i7}{Supervisory responsibility}
 \item{i8}{Belong to a trade union or staff associa}
 \item{i9}{Self assessed social class}
 \item{j1}{Sex}
 \item{j2}{Year of Birth}
 \item{j3own}{Country of Birth}
 \item{j3mum}{Mothers Country of Birth}
 \item{j3dad}{Fathers Country of Birth}
 \item{j4}{Age R came to Australia}
 \item{j5}{Fathers Occupation at R Age 14}
 \item{j6dad}{Fathers political preference at R Age 14}
 \item{j6mum}{Mothers political preference at R Age 14}
 \item{j7}{Religion}
 \item{j8}{Religious attendance}
 \item{j9}{Marital Status}
 \item{j10}{Occupation of Spouse}
 \item{j11}{Occupation of Spouse}
 \item{j12}{Employer of Spouse}
 \item{j13}{Does Spouse supervise people at work}
 \item{j14}{Spouse member of a trade union}
 \item{j15}{Spouses political identification}
 \item{d3spend1}{Spending on Education}
 \item{d3spend2}{Spending on Defence}
 \item{d3spend3}{Spending on Aged Pensions}
 \item{d3spend4}{Spending on Health}
 \item{d3spend5}{Spending on Roads}
 \item{d3spend6}{Spending on Police}
 \item{d3spend7}{Spending on Unemployment Benefits}
 \item{d3spend8}{Spending on Help for industry}
 \item{d3spend9}{Spending on Protecting the Environment}
 \item{postcode}{Postcode of Residence}
 \item{elecdiv}{Electoral Division}
 \item{state}{State of Residence}
 \item{statemap}{State mapping variable}
}
}

\keyword{datasets}
