% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes_1996.R
\docType{data}
\name{aes_1996}
\alias{aes_1996}
\title{Australian Election Study 1996}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1797 rows and 296 columns.}
\source{
\url{http://www.australianelectionstudy.org/voter_studies.html}
}
\usage{
aes_1996
}
\description{
\itemize{
   \item Total sample: 3,000
   \item Valid responses: 1,795
   \item Effective response rate: 61.8
}
}
\details{
The 1996 Australian Election Study is the fourth in a series of surveys beginning in 1987 which
have been timed to coincide with Australian Federal elections. The series also builds on the 1967
and 1979 Australian Political Attitudes Surveys. The Australian Election Studies aim to provide a
long-term perspective on stability and change in the political attitudes and behaviour of the
Australian electorate, and investigate the changing social bases of Australian politics as the
economy and society modernise and change character. In addition to these long-term goals they
examine the political issues prevalent in the election and assesses their importance for the
election result.

The 1996 survey replicates many questions from the 1987, 1990 and 1993 Australian Election
Studies, but also introduces a variety of new questions including a section on national identity.
Other sections covered the respondent's interest in the election campaign and politics, their
past and present political affiliation, evaluation of parties and candidates, alignment with
parties on various election issues, evaluation of the economic situation and economic policies,
attitudes to a range of environmental issues, attitudes to foreign affairs and defence issues and
attitudes to contemporary social policy issues including equal opportunity, censorship,
migration, assistance for aborigines, abortion, criminal law, expenditure on social services, the
monarchy and the Australian flag.

Background variables include level of education, employment status, occupation, type of employer,
position at workplace, trade union membership, sex, age, own and parents country of birth,
parents' political preferences, religion, marital status, income, and where applicable, the
occupation, trade union membership and political preference of the respondent's spouse.

\strong{Useful Links}
\itemize{
   \item \href{http://ada.edu.au/ADAData/codebooks/ADA.CODEBOOK.00943.pdf.zip}{Original codebook}
   \item \href{http://ada.edu.au/ADAData/questionnaires/ADA.QUESTIONNAIRE.00763.pdf.zip}{Original questionnaire}
}
}
\section{Variables}{


Note: this documentation has been autogenerated from the dataset labels not the original codebook.

\describe{
  \item{caseid}{Case identification number}
  \item{a1}{Interest in politics}
  \item{a2}{Followed election news in the newspapers}
  \item{a3p1}{Followed election news on TV}
  \item{a3p2}{Followed election news on the radio}
  \item{a4}{Interest in election campaign}
  \item{a5}{Watched televised debates}
  \item{a6}{Who performed better in televised debates}
  \item{a7p1}{During campaign: discussed politics w others}
  \item{a7p2}{During campaign: talk to people about vote}
  \item{a7p3}{During campaign: go to meetings or rallies}
  \item{a7p4}{During campaign: contribute money}
  \item{a7p5}{During campaign: work for party/candidate}
  \item{a8p1}{Currently a party member}
  \item{a8p2}{Ever been a party member}
  \item{a9}{Contact with Federal politician}
  \item{a10}{Should be compulsory voting}
  \item{a11}{Would vote if not compulsory}
  \item{a12}{Satisfied with Aust. democracy}
  \item{b1}{Party identification}
  \item{b2}{Strength of party identification}
  \item{b3}{When decided which way to vote}
  \item{b4}{Care which party wins election}
  \item{b5}{Most important in vote decision}
  \item{b6}{Difference between Labor and Liberal}
  \item{b7}{Parties doing good or bad job}
  \item{b8own}{Own left-right position}
  \item{b8lib}{Liberal left-right position}
  \item{b8alp}{Labor left-right posiiton}
  \item{b8nat}{National left-right position}
  \item{b8demo}{Democrats left-right position}
  \item{b8green}{Greens left-right position}
  \item{b9reps}{Vote in House of Representatives}
  \item{b9senate}{Vote in Senate}
  \item{b10reps}{Preferences to in House of Reps}
  \item{b10senat}{Preferences to in Senate}
  \item{b11}{Follow How to Vote card for H of R}
  \item{b12}{Party box or own prefs for Senate}
  \item{b13}{Thought of changing vote during campaign}
  \item{b14}{Vote in House of Representatives in 1993}
  \item{b15p1}{Previous vote - always same}
  \item{b15p2}{Previous vote - sometimes Liberal}
  \item{b15p3}{Previous vote - sometimes Labor}
  \item{b15p4}{Previous vote - sometimes National}
  \item{b15p5}{Previous vote - sometimes Democrat}
  \item{b15p6}{Previous vote - sometimes others}
  \item{b15p7}{Previous vote - not before this election}
  \item{b16p1}{Political parties care what people think}
  \item{b16p2}{Political parties make system work}
  \item{b16p3}{Makes a difference who is in power}
  \item{b17lib}{Feelings about Liberal Party}
  \item{b17alp}{Feelings about Labor Party}
  \item{b17nat}{Feelings about National Party}
  \item{b17demo}{Feelings about Australian Democrats}
  \item{b17green}{Feelings about Greens}
  \item{c1keat}{Feelings about Paul Keating}
  \item{c1how}{Feelings about John Howard}
  \item{c1fisch}{Feelings about Tim Fischer}
  \item{c1kern}{Feelings about Cheryl Kernot}
  \item{c2p1}{Federal politicians improper}
  \item{c2p2}{Federal politicians high morals}
  \item{c2p3}{Federal politicians tell lies}
  \item{c2p4}{Federal politicians special interests}
  \item{c2p5}{Federal politicians low ethics}
  \item{c3pkmor}{Paul Keating moral}
  \item{c3pkint}{Paul Keating intelligent}
  \item{c3pkcom}{Paul Keating compassionate}
  \item{c3pksen}{Paul Keating sensible}
  \item{c3pklea}{Paul Keating provides strong leadership}
  \item{c3pkhon}{Paul Keating honest}
  \item{c3pkrel}{Paul Keating reliable}
  \item{c3pkkno}{Paul Keating knowledgeable}
  \item{c3pkins}{Paul Keating inspiring}
  \item{c3pkdep}{Paul Keating dependable}
  \item{c3pkarr}{Paul Keating arrogant}
  \item{c3jhmor}{John Howard moral}
  \item{c3jhint}{John Howard intelligent}
  \item{c3jhcom}{John Howard compassionate}
  \item{c3jhsen}{John Howard sensible}
  \item{c3jhlea}{John Howard provides strong leadership}
  \item{c3jhhon}{John Howard honest}
  \item{c3jhrel}{John Howard reliable}
  \item{c3jhkno}{John Howard knowledgeable}
  \item{c3jhins}{John Howard inspiring}
  \item{c3jhdep}{John Howard dependable}
  \item{c3jharr}{John Howard arrogant}
  \item{c4name}{Name of local Federal MP}
  \item{c4party}{Party of local Federal MP}
  \item{c5}{Government look after themselves}
  \item{c6p1}{Fed Politician act honestly}
  \item{c6p2}{Fed Politician act for public interest}
  \item{c6p3}{Fed Politician be dignified}
  \item{c6p4}{Fed Politician refuse gifts}
  \item{c6p5}{Fed Politician no special interest}
  \item{c6p6}{Fed Politician tell truth}
  \item{c6p7}{Fed Politician use resources well}
  \item{c6p8}{Fed Politician respect public}
  \item{c7p1}{Fed Pol know what people think}
  \item{c7p2}{Doesnt make a difference who vote for}
  \item{d1tax}{Party closest to own view taxation}
  \item{d1immig}{Party closest to own view immigration}
  \item{d1educ}{Party closest to own view education}
  \item{d1envir}{Party closest to own view environment}
  \item{d1indrel}{Party closest to own view industrial relations}
  \item{d1health}{Party closest to own view on health and Medicare}
  \item{d1asia}{Party closest to own view on Asia}
  \item{d1def}{Party closest to own view on defence}
  \item{d1intrat}{Party closest to own view interest rates}
  \item{d1unemp}{Party closest to own view unemployment}
  \item{d1priv}{Party closest to own view privatisation}
  \item{d1inflat}{Party closest to own view on inflation}
  \item{d1stater}{Party closest to own view State,Ter. issues}
  \item{d2tax}{Importance of taxation}
  \item{d2immig}{Importance of immigration}
  \item{d2educ}{Importance of education}
  \item{d2envir}{Importance of environment}
  \item{d2indrel}{Importance of industrial relations}
  \item{d2health}{Importance of health and Medicare}
  \item{d2asia}{Importance of links with Asia}
  \item{d2def}{Importance of defence}
  \item{d2intrat}{Importance of interest rates}
  \item{d2unemp}{Importance of unemployment}
  \item{d2priv}{Importance of privatisation}
  \item{d2inflat}{Importance of inflation}
  \item{d2stater}{Importance of State,Ter. issues}
  \item{d3p1}{Issue of most concern last 12 months}
  \item{d3p2}{Second issue of concern last 12 months}
  \item{d4}{Have private health insurance}
  \item{d5hhold}{Financial situation of hhold cf 1 yr ago}
  \item{d5cntry}{Economic situation country cf 1 yr ago}
  \item{d6hhold}{Effect of govt policies on hhold situation}
  \item{d6cntry}{Effect of govt policies on country situation}
  \item{d7hhold}{Financial situation of hhold cf in 1 yr}
  \item{d7cntry}{Economic situation country cf in 1 yr}
  \item{d8hhold}{Effect of govt policies on hhold situation in 1yr}
  \item{d8cntry}{Effect of govt policies on country situation in 1yr}
  \item{d9tax}{High tax makes people less willing to work}
  \item{d9tupow}{Trade unions have too much power}
  \item{d9bigpow}{Big business has too much power}
  \item{d9notu}{Aust better off without trade unions}
  \item{d9equal}{Income & wealth should be redistributed}
  \item{d9strict}{Stricter laws to regulate trade unions}
  \item{d9empcon}{Individual contracts better than enterprise agreements}
  \item{d9ecgrow}{Faster economic growth more than income distribution}
  \item{d9chostu}{Rather be in a Trade Union than not}
  \item{e1}{Choice between taxes and social services}
  \item{e2migeo}{Equal opportunities for migrants}
  \item{e2nudsex}{Nudity & sex in films & magazines}
  \item{e2abland}{Aboriginal land rights}
  \item{e2equop}{Equal opportunities for women}
  \item{e2immig}{No. of migrants allowed into Australia}
  \item{e2abor}{Government help for aborigines}
  \item{e2asia}{Building closer links with Asia}
  \item{e3}{Abortion}
  \item{e5deathp}{Death penalty reintroduced for murder}
  \item{e5marij}{Decriminalise smoking of marijuana}
  \item{e5ration}{Policies based on rational calculation}
  \item{e5lawbrk}{Stiffer sentences if break law}
  \item{e5pref}{Women given preferential treatment}
  \item{e5succes}{In Australian society anyone can succeed}
  \item{e5opp}{Increase business opportunities for women}
  \item{e6nengy}{Nuclear energy a necessity}
  \item{e6presn}{Nature most precious in life}
  \item{e6uran}{Australian should mine its uranium}
  \item{e6spend}{Increase spending to protect environment}
  \item{e6bias}{TV is biased in favour of greens}
  \item{e6pollut}{Stronger measures against pollution}
  \item{e7pollut}{Pollution urgent envir. concern}
  \item{e7pop}{Overpopulation urgent envir. concern}
  \item{e7waste}{Waste disposal urgent envir. concern}
  \item{e7uran}{Uranium mining urgent envir. concern}
  \item{e7log}{Logging urgent envir. concern}
  \item{e7wild}{Wildlife urgent envir. concern}
  \item{e7soil}{Soil urgent envir. concern}
  \item{e7ghouse}{Greenhouse urgent envir. concern}
  \item{e7urg1}{Most urgent envir. concern}
  \item{e7urg2}{Second most urgent envir. concern}
  \item{e8envir_}{Approve environmental groups}
  \item{e8uran}{Approve anti-uranium mining groups}
  \item{e8war}{Approve anti-war and anti-nuclear groups}
  \item{e9}{Member of environment group}
  \item{e10}{Views on prices and the environment}
  \item{e11}{Worried about environment cf 5 years ago}
  \item{e12}{Views on taxes and the environment}
  \item{e13aims1}{Aims of Australia - first choice}
  \item{e13aims2}{Aims of Australia - second choice}
  \item{f1}{Government spend more on defence}
  \item{f2japan}{Japan threat to Aust. security}
  \item{f2us}{United States threat to Aust. security}
  \item{f2china}{China threat to Aust. security}
  \item{f2png}{Papua New Guinea threat to Aust. security}
  \item{f2vietn}{Vietnam threat to Aust. security}
  \item{f2malay}{Malaysia threat to Aust. security}
  \item{f2singap}{Singapore threat to Aust. security}
  \item{f2indon}{Indonesia threat to Aust. security}
  \item{f2nz}{New Zealand threat to Aust. security}
  \item{f2india}{India threat to Aust. security}
  \item{f3japan}{Japan important defence relations}
  \item{f3us}{United States important defence relations}
  \item{f3china}{China important defence relations}
  \item{f3png}{Papua New Guinea important defence relations}
  \item{f3vietn}{Vietnam important defence relations}
  \item{f3malay}{Malaysia important defence relations}
  \item{f3singap}{Singapore important defence relations}
  \item{f3indon}{Indonesia important defence relations}
  \item{f3nz}{New Zealand important defence relations}
  \item{f3india}{India important defence relations}
  \item{f4}{United States alliance under ANZUS}
  \item{f5}{Benefits from Australian alliance with US}
  \item{f6p1}{Aust defence stronger than 10 yrs ago}
  \item{f6p2}{Security agreement between Aust and Indon}
  \item{f6p3}{Pacific region less safe}
  \item{f6p4}{Aust would be able to defend itself}
  \item{f6p5}{Should organize defence with Asia}
  \item{f7p1}{Aust trade lies with Asia}
  \item{f7p2}{Aust use tariffs to protect industry}
  \item{f7p3}{Aust should side with Japan not US}
  \item{f7p4}{Japan influence too great in Aust}
  \item{g1bnaust}{Truly Australian be born in Aust}
  \item{g1auscit}{Truly Australian be citizen}
  \item{g1livaus}{Truly Australian live in Aust}
  \item{g1spkeng}{Truly Australian speak English}
  \item{g1christ}{Truly Australian be Christian}
  \item{g1auslaw}{Truly Australian respect law}
  \item{g1feelau}{Truly Australian feel Aust}
  \item{g2democ}{Proud of Australian democracy}
  \item{g2polinf}{Proud of Australian influence}
  \item{g2econom}{Proud of Australian ecenomics}
  \item{g2socsec}{Proud of Australian security}
  \item{g2scienc}{Proud of Australian science}
  \item{g2histry}{Proud of Australian history}
  \item{g2forces}{Proud of Australian forces}
  \item{g2sport}{Proud of Australian sports}
  \item{g2arts}{Proud of Australian art}
  \item{g3}{Importance of Queen}
  \item{g4}{Australia a republic}
  \item{g5}{Head of republic from voters}
  \item{g6}{Change or retain Aust flag}
  \item{g7p1}{Immigrants increase crime}
  \item{g7p2}{Immigrants good for economy}
  \item{g7p3}{Immigrants take jobs from Aust}
  \item{g7p4}{Immigrants make Aust more open}
  \item{g8}{Number of immigrants increased}
  \item{g9}{Permanent residents be citizens}
  \item{g10p1}{Limit imports}
  \item{g10p2}{Schools teach foreign languages}
  \item{g10p3}{Aust follow own interests regardless}
  \item{g10p4}{Foreigners shouldnt buy land}
  \item{g11p1}{Rather be Aust citizen than else}
  \item{g11p2}{Proud of Aust sports performances}
  \item{g11p3}{World better if more like Aust}
  \item{g11p4}{Aust better than other counties}
  \item{g12p1}{Quiz Howard leader of Liberals}
  \item{g12p2}{Quiz 75 in House of Reps}
  \item{g12p3}{Quiz Aust a federation in 1901}
  \item{g12p4}{Quiz High Court change Constitution}
  \item{g12p5}{Quiz Senators not in Cabinet}
  \item{g12p6}{Quiz Pay to stand for Parliament}
  \item{g12p7}{Quiz Four years between elections}
  \item{g12p8}{Quiz Parliament before sign treaty}
  \item{g12p9}{Quiz Senate proportional rep}
  \item{h1}{Age left school}
  \item{h2}{Highest qual obtained since leaving school}
  \item{h3}{No. years of tertiary education}
  \item{h4}{Occupation last week}
  \item{h5occ}{Occupation}
  \item{h5pos}{Position in organisation}
  \item{h5employ}{Employer}
  \item{h5superv}{Supervisory responsibility}
  \item{h6}{Belong to a trade union or staff association}
  \item{h7}{Can choose trade union membership}
  \item{h8now}{How are wages and conditions set}
  \item{h8pref}{How prefer wages and conditions set}
  \item{j1}{Sex}
  \item{j2}{Year of birth}
  \item{j3own}{Country of birth}
  \item{j3mum}{Mothers country of birth}
  \item{j3dad}{Fathers country of birth}
  \item{j4dad}{Fathers political preference at R age 14}
  \item{j4mum}{Mothers political preference at R age 14}
  \item{j5}{Religion}
  \item{j6}{Religious attendance}
  \item{j7}{Self assessed social class}
  \item{j8}{Marital status}
  \item{j9}{Own, buying or rent home}
  \item{j10}{Coping with mortgage repayments}
  \item{j11}{Live rural or urban}
  \item{j12}{Gross annual income}
  \item{j13}{Spouse working}
  \item{j14}{Spouse occupation}
  \item{j15}{Spouse employer}
  \item{j16}{Does spouse supervise people at work}
  \item{j17}{Spouse member of trade union}
  \item{j18}{Spouses political identification}
  \item{retwk}{Week of return}
  \item{elect}{Electoral Division}
}
}

\keyword{datasets}
