% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.palletes.R
\docType{data}
\name{bluepurp}
\alias{bluepurp}
\title{Color palletes to make the heart sing lol
Herein lies the warm tones to make your plot pop.
"@color.pallete" cinema on instagram is the main source of the movie specific palletes
To see the individual colors and plot them , see the R script in /exec
bluepurp}
\format{
a list of five bluish purple colors
\describe{
}
}
\usage{
bluepurp
}
\description{
Color palletes to make the heart sing lol
Herein lies the warm tones to make your plot pop.
"@color.pallete" cinema on instagram is the main source of the movie specific palletes
To see the individual colors and plot them , see the R script in /exec
bluepurp
}
\keyword{datasets}
