% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.palletes.R
\docType{data}
\name{bluepurp}
\alias{bluepurp}
\alias{earthy.blueyellow}
\alias{earthy.forest}
\alias{get.out}
\alias{handmaids}
\alias{rando}
\alias{sorry2botheryou}
\alias{taxi.driver}
\alias{itv.pv}
\alias{earth_cb}
\alias{earth_teal_ochre}
\alias{.woodcut_palettes}
\title{earthy.blueyellow}
\format{
Lists of color palletes stored as objects

\describe{
\item{bluepurp}{a list of five bluish purple colors}
\item{earthy.blueyellow}{a list of five earthy bluish yellow colors }
\item{earthy.forest}{a list of 5 earthy greens}
\item{get.out}{a list of 7 colors based on the movie get out }
\item{handmaids}{a list of 7 colors based on the show handmaids tale}
\item{rando}{a list of 10 random earthy colors that I like}
\item{sorry2botheryou}{a list of 7 colors based on the movie Sorry to bother you}
\item{taxi.driver}{a list of 6 colors based on the movie Taxi Driver}
}

An object of class \code{character} of length 5.

An object of class \code{character} of length 5.

An object of class \code{character} of length 7.

An object of class \code{character} of length 7.

An object of class \code{character} of length 10.

An object of class \code{character} of length 7.

An object of class \code{character} of length 6.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.

An object of class \code{list} of length 11.
}
\usage{
bluepurp

earthy.blueyellow

earthy.forest

get.out

handmaids

rando

sorry2botheryou

taxi.driver

itv.pv

earth_cb

earth_teal_ochre

.woodcut_palettes
}
\description{
All of the names of color palletes in woodcut package.
}
\details{
Color paletes to make the heart sing lol. Herein lies the warm tones to make your plot pop.

I derived a lot of the movie specific palettes from the "@color.pallete.cinema" page on Instagram.

To see the individual colors and plot them see \code{woodcut::woodcut_palettes()}.
}
\keyword{bluepurp}
\keyword{colors}
\keyword{datasets}
