% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.palletes.R
\docType{data}
\name{bluepurp}
\alias{bluepurp}
\alias{earthy.blueyellow}
\alias{earthy.forest}
\alias{get.out}
\alias{handmaids}
\alias{rando}
\alias{sorry2botheryou}
\alias{taxi.driver}
\title{Color palletes to make the heart sing lol
Herein lies the warm tones to make your plot pop.
"@color.pallete" cinema on instagram is the main source of the movie specific palletes
To see the individual colors and plot them , see the R script in /exec}
\format{
Lists of color palletes stored as objects
\describe{
\item{bluepurp}{a list of five bluish purple colors}
\item{earthy.blueyellow}{a list of five earthy bluish yellow colors }
\item{earthy.forest}{a list of 5 earthy greens}
\item{get.out}{a list of 7 colors based on the movie get out }
\item{handmaids}{a list of 7 colors based on the show handmaids tale}
\item{rando}{a list of 10 random earthy colors that I like}
\item{sorry2botheryou}{a list of 7 colors based on the movie Sorry to bother you}
\item{taxi.driver}{a list of 6 colors based on the movie Taxi Driver}
}

An object of class \code{character} of length 5.

An object of class \code{character} of length 5.

An object of class \code{character} of length 7.

An object of class \code{character} of length 7.

An object of class \code{character} of length 10.

An object of class \code{character} of length 7.

An object of class \code{character} of length 6.
}
\usage{
bluepurp

earthy.blueyellow

earthy.forest

get.out

handmaids

rando

sorry2botheryou

taxi.driver
}
\description{
Color palletes to make the heart sing lol
Herein lies the warm tones to make your plot pop.
"@color.pallete" cinema on instagram is the main source of the movie specific palletes
To see the individual colors and plot them , see the R script in /exec

earthy.blueyellow

earthy.forest

Get Out

Handmaids Tale

Randos

Sorry to bother you

Taxi Driver
}
\keyword{bluepurp}
\keyword{colors}
\keyword{datasets}
