% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_colpal.R
\name{colpal_proj}
\alias{colpal_proj}
\title{Construct color palette for continuous or discrete data}
\usage{
colpal_proj(cols = c("earth", "soft"), type = c("discrete", "continuous"))
}
\arguments{
\item{cols}{Earth tones, soft tones, etc.}

\item{type}{Discrete or continuous}
}
\value{
Color palette vector
}
\description{
Construct color palette for continuous or discrete data
}
