% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.palletes.R
\docType{data}
\name{earthy.blueyellow}
\alias{earthy.blueyellow}
\title{earthy.blueyellow}
\format{
a list of five earthy bluish yellow colors
\describe{
}
}
\usage{
earthy.blueyellow
}
\description{
earthy.blueyellow
}
\keyword{datasets}
