% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_colpal.R
\name{fetch_pal}
\alias{fetch_pal}
\title{Fetch color palettes}
\usage{
fetch_pal(
  palette = c("bluepurp", "earthyby", "forest", "get.out", "handmaids", "rando",
    "sorry2botheryou", "taxi.driver", "itv.pv"),
  reverse = FALSE
)
}
\arguments{
\item{palette}{Palette name}

\item{reverse}{Reverse the order of colors}
}
\value{
A function that takes an integer n and returns n colors from the specified palette
}
\description{
Fetch color palettes
}
