% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_docs.R
\name{get_names_and_labels}
\alias{get_names_and_labels}
\title{Get the names and labels of a labelled dataframe}
\usage{
get_names_and_labels(df)
}
\arguments{
\item{df}{a labelled dataframe}
}
\value{
A dataframe containing the names and labels of the dataset
}
\description{
Get the names and labels of a labelled dataframe
}
