% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.palletes.R
\docType{data}
\name{handmaids}
\alias{handmaids}
\title{Handmaids Tale}
\format{
a list of 7 colors based on the show handmaids tale
\describe{
}
}
\usage{
handmaids
}
\description{
Handmaids Tale
}
\keyword{datasets}
