% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_docs.R
\name{make_docs}
\alias{make_docs}
\title{Autogenerate documentation from the AES survey response dataset}
\usage{
make_docs(df)
}
\arguments{
\item{df}{A labelled AES dataset}
}
\description{
Autogenerate documentation from the AES survey response dataset
}
