% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotaxes.R
\name{plotaxis}
\alias{plotaxis}
\title{Plot axis creation}
\usage{
plotaxis(dat, colName, round = 2)
}
\arguments{
\item{dat}{A dataframe with variable to create axis information for}

\item{colName}{Name of variable do you need to find intervals for (make sure there are quotations around the colnames )}

\item{round}{Number of digits to round axis values to. Default is 2}
}
\value{
A list of three values
}
\description{
The purpose of this function is to create  even plot ticks at the minimum, middle and max values of the plot.
}
\examples{
#create three evenly spaced axis ticks for weight column on mtcars dataset
axis_ticks<-plotaxis(dat=mtcars,colName = "wt")

axis_ticks

#Slightly more complicated 
#create three evenly spaced axis ticks for airquality dataset
months<-unique(airquality$Month)#unique list of months of year sampled

dats<-lapply(months,#create list of datasets, one for each month
 function (month) dplyr::filter(airquality,Month==month))
axis_ticks<-lapply(dats, #create set of evenly spaced axis tick values for all datasets.
function(x) plotaxis(dat = x, colName = "Ozone"))

axis_ticks 
}
\concept{aesthetics}
\keyword{aesthetics}
\keyword{plotting}
