% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_colpal.R
\name{scale_color_bean}
\alias{scale_color_bean}
\alias{scale_fill_bean}
\title{Discrete color palettes for ggplot2 figures}
\usage{
scale_color_bean(
  palette = "earthyby",
  discrete = TRUE,
  reverse = FALSE,
  na.value = "grey80",
  ...
)

scale_fill_bean(
  palette = "earthyby",
  discrete = TRUE,
  reverse = FALSE,
  na.value = "grey80",
  ...
)
}
\arguments{
\item{palette}{Palette name}

\item{discrete}{Boolean for discrete or continuous}

\item{reverse}{Reverse the order of colors}

\item{na.value}{Color for NA values}

\item{...}{Additional arguments passed to ggplot2 functions}
}
\value{
Scale functions for ggplot2

Scale functions for ggplot2
}
\description{
Discrete color palettes for ggplot2 figures

Discrete fill palettes for ggplot2 figures
}
