% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_colpal.R
\name{scale_color_project}
\alias{scale_color_project}
\title{Color scale (Continuous or Discrete) for project}
\usage{
scale_color_project(
  cols = c("earth", "soft"),
  type = c("discrete", "continuous"),
  n = NULL
)
}
\arguments{
\item{cols}{Earth tones, soft tones, etc.}

\item{type}{Continuous or discrete}

\item{n}{How many colors to return (only for continuous)}
}
\value{
Color scale function for ggplot2
}
\description{
Color scale (Continuous or Discrete) for project
}
