% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.palletes.R
\docType{data}
\name{sorry2botheryou}
\alias{sorry2botheryou}
\title{Sorry to bother you}
\format{
a list of 7 colors based on the movie Sorry to bother you
\describe{
}
}
\usage{
sorry2botheryou
}
\description{
Sorry to bother you
}
\keyword{datasets}
