% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.palletes.R
\docType{data}
\name{taxi.driver}
\alias{taxi.driver}
\title{Taxi Driver}
\format{
a list of 6 colors based on the movie Taxi Driver
\describe{
}
}
\usage{
taxi.driver
}
\description{
Taxi Driver
}
\keyword{datasets}
