% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_woodcut.R
\name{theme_woodcut}
\alias{theme_woodcut}
\title{Base plotting theme}
\usage{
theme_woodcut(
  font_family = "Helvetica Neue",
  font_size = 24,
  base_color = "#3f3f3f",
  axis_line_color = "black",
  panel_border_color = "black",
  panel_fill = "white",
  facet_strip_fill = "white",
  legend_key_height = unit(3, "cm")
)
}
\arguments{
\item{font_family}{Font type}

\item{font_size}{General font size}

\item{base_color}{General color of titles and labels}

\item{axis_line_color}{Color of axis lines}

\item{panel_border_color}{Color of panels}

\item{panel_fill}{Color filling the panels}

\item{facet_strip_fill}{Color of the panel facet labels}

\item{legend_key_height}{Height of the legend with plot}
}
\value{
Themed ggplot
}
\description{
Base plotting theme
}
