% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woodcut_palettes.R
\name{woodcut_palettes}
\alias{woodcut_palettes}
\title{Palettes provided in woodcut package}
\usage{
woodcut_palettes(names = NULL, plot = interactive())
}
\arguments{
\item{names}{Optional character vector of palette names to return.
If NULL (default), returns all.}

\item{plot}{If TRUE, draw swatches (default: interactive())}
}
\value{
Color palette options either as a tibble or list (default) and a plot of the colors (optional)
}
\description{
List woodcut palettes (and optionally preview)
}
